Re-organized security to user module
