+++ backend/src/timestamp/service.py
@@ -1,6 +1,6 @@
 import timestamp.repository as repository
 from dbutil import DBTransactionDependency
-from routers.security import LightWeightUser, get_current_user
+from user.service import LightWeightUser, get_current_user
 from typing import Annotated, List
 from fastapi import Depends, Form, APIRouter
