+++ backend/src/user/router.py
@@ -1,0 +1,2 @@
+# Re-expose router object.
+from user.service import router
